requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
cryptography>=42.0.0
python-multipart>=0.0.9
websockets>=12.0
pymupdf>=1.23.0
python-docx>=1.1.0
pillow>=10.0.0
